refactor: unify session state initialization in demo-oauth2
- Move session state initialization to demo-oauth2 binary
- Rename init_oauth2_state to oauth2_state_init for consistency
- Remove duplicate session state initialization from liboauth2
- Rename oauth2/engine.rs to oauth2/core.rs for better clarity
- Update imports and function references accordingly

This change ensures session state is initialized only once at the top level,
preparing for future integration with libpasskey.
