+++ liboauth2/src/oauth2/mod.rs
@@ -1,7 +1,7 @@
-mod engine;
+mod core;
 mod idtoken;
 
-pub(crate) use engine::{
+pub(crate) use core::{
     csrf_checks, get_user_oidc_oauth2, prepare_oauth2_auth_request, validate_origin,
 };
 pub(crate) use idtoken::TokenVerificationError;
